offline cart

+++ src/components/CartItem.js
@@ -2,24 +2,70 @@
 /* eslint-disable jsx-a11y/img-redundant-alt */
 import React from 'react'
 import axios from 'axios'
+import Cookies from 'universal-cookie'
+
 import 'bootstrap/dist/css/bootstrap.min.css'
+
+const cookies = new Cookies()
 
 class CartItem extends React.Component {
 
     onIncreaseClick = () => {
-        axios.get(`${process.env.REACT_APP_API_URL}/product/${this.props.item._id}`).then((res) => {
+        if (cookies.get('token')) {
+            axios.get(`${process.env.REACT_APP_API_URL}/product/${this.props.item._id}`).then((res) => {
+                alert('Ürün sepete eklendi (1)')
+            }).catch((err) => {
+                alert(err.response.data.error)
+            })
+        } else {
+            const cart = window.localStorage.getItem('cart')
+
+            if (cart) {
+                const cartAsArray = JSON.parse(cart)
+                const foundProduct = cartAsArray.find((cartProduct) => cartProduct._id === this.props.item._id)
+                if (foundProduct) {
+                    cartAsArray[cartAsArray.indexOf(foundProduct)].quantity++
+                } else {
+                    cartAsArray.push({ _id: this.props.item._id, quantity: 1 })
+                }
+                window.localStorage.setItem('cart', JSON.stringify(cartAsArray))
+            } else {
+                window.localStorage.setItem('cart', JSON.stringify([{ _id: this.props.item._id, quantity: 1 }]))
+            }
+
             alert('Ürün sepete eklendi (1)')
-        }).catch((err) => {
-            alert(err.response.data.error)
-        })
+        }
     }
 
     onDecreaseClick = () => {
-        axios.delete(`${process.env.REACT_APP_API_URL}/product/${this.props.item._id}`).then((res) => {
-            alert('Ürün sepetten çıkarıldı (1)')
-        }).catch((err) => {
-            alert(err.response.data.error)
-        })
+        if (cookies.get('token')) {
+            axios.delete(`${process.env.REACT_APP_API_URL}/product/${this.props.item._id}`).then((res) => {
+                alert('Ürün sepetten çıkarıldı (1)')
+            }).catch((err) => {
+                alert(err.response.data.error)
+            })
+        } else {
+            const cart = window.localStorage.getItem('cart')
+
+            if (cart) {
+                const cartAsArray = JSON.parse(cart)
+                const foundProduct = cartAsArray.find((cartProduct) => cartProduct._id === this.props.item._id)
+                if (foundProduct) {
+                    cartAsArray[cartAsArray.indexOf(foundProduct)].quantity--
+
+                    if (cartAsArray[cartAsArray.indexOf(foundProduct)].quantity === 0)
+                        cartAsArray.splice(cartAsArray.indexOf(foundProduct), 1)
+
+
+                    alert('Ürün sepetten çıkarıldı (1)')
+                } else {
+                    alert('Ürün sepette bulunamadı')
+                }
+                window.localStorage.setItem('cart', JSON.stringify(cartAsArray))
+            } else {
+                alert('Ürün sepette bulunamadı')
+            }
+        }
     }
 
     render() {
